Fix: TreePickerNode Spinner on async/large lists expandNode
An issue where loading large subtrees (> 250 nodes) would fail to render
the spinner. State was being correctly updated but the view wasn't being
redrawn. React was bundling the redraw of the spinner with the redraw of
the loading of the new subtree.

 - Add interval to defer rendering the list by 25ms which will also show
a spinner for times when the users' machine is running slow (not just
async calls)
 - Refactor code slightly (made fixing the bug clearer and component
simpler)
 - Moved getExpander to it's own component without the need for a func